Back_end/requirements.txt
#软件包列表： 软件包名==版本

fastapi==0.124.4
uvicorn==0.38.0
sqlalchemy==2.0.45
python-jose[cryptography]==3.5.0
passlib==1.7.4
python-multipart==0.0.20
pydantic==2.12.5
pydantic-core==2.41.5
dashscope==1.25.5
pymysql==1.1.2
mysql-connector-python==9.1.0
requests==2.32.5
cryptography==46.0.3
scikit-learn==1.8.0
numpy==2.4.3
scipy==1.17.1
aiohttp==3.13.2
aiosqlite==0.22.0
alembic==1.17.2
databases==0.9.0
tortoise-orm==0.25.2